Introduce the base for 'treatment' (pakkeforløp)
Note that this commit is just the first of many, so this is work in progress.

- Introduce 'clock' svg, imported from Sketch
- Introduce 'IconWithText' component
- Introduce 'TreatmentsDuration' component
- Introduce placeholders for all page variations we're going to refer to
- Introduce 'uuid' package to make it easier to define unique ids
- Expand text alignment classes

We have to find out if the styles in Sketch are to be applied as they are, but I assume most of them are 'off' and should be ignored as we want to remain consistent with the existing styleguide.