Introduce two more types of buttons: 'optionPrimary' and 'optionSecondary'
This is necessary if we are to implement the UX mockup as it is. The buttons are represented at the top of the 'Lungekreft' page. I also did:

- Update the documentation of 'Button'
- Simplified 'Button' itself
- Made it possible for it to receive other classes without crashing

Now we can pass classes like spacing classses (ie. 'l-mt-4') directly to the button while keeping the classes that were passed as boolean props.