Bring 'createUniqueHeaders' logic back to 'SidebarSection' component, as we must know if 'props.list' has been passed to it or not
This limitation was there by design before we attempted to generate unique ids to ALL 'h2' headings, and it was thete for good reason.

If 'list' is passed as 'props' to the 'SidebarSection' component (as a data attribute), the backend takes over and the 'id' generation must not proceeed. This could be one of the causes why links have stopped to work on certain pages, but it would have to be confirmed.